requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via pytest
beautifulsoup4==4.11.1
    # via -r requirements.in
black==22.8.0
    # via -r requirements.in
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
factory-boy==3.2.1
    # via -r requirements.in
faker==15.0.0
    # via factory-boy
flake8==5.0.4
    # via -r requirements.in
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy==0.971
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via pytest
pathspec==0.10.1
    # via black
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pytest
types-beautifulsoup4==4.11.6
    # via -r requirements.in
types-requests==2.28.11
    # via -r requirements.in
types-urllib3==1.26.24
    # via types-requests
typing-extensions==4.3.0
    # via mypy
urllib3==1.26.12
    # via requests
